feat: move logger to its own file
